+++ package.json
@@ -43,6 +43,7 @@
     "framer-motion": "^11.0.18",
     "lucide-react": "^0.469.0",
     "motion": "^11.15.0",
+    "nuqs": "^2.4.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-helmet-async": "^2.0.5",
