+++ src/components/ui/index.ts
@@ -6,4 +6,5 @@ export * from "./circle-progress";
 export * from "./badge";
 export * from "./custom-progress";
 export * from "./carousel";
-export * from "./chart"
+export * from "./chart";
+export * from "./modal";
