+++ src/layouts/modal-layout/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-group-content"
